requirements.txt
aiohttp==3.10.10
APScheduler==3.10.4
chromadb==0.5.13
fastapi==0.115.3
firebase_admin==6.5.0
langchain==0.3.4
langchain_community==0.3.3
langchain_core==0.3.12
langchain_openai==0.2.3
openai==1.52.1
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.1.1
Requests==2.32.3
scikit_learn==1.5.2
uvicorn==0.32.0
python-multipart==0.0.12
coolsms==0.4
coolsms_python_sdk==2.0.3